It needs some fixing to do, I just winged it.
Generated an Announcement Module. created service and controller file that has functionality of deleting, creating, finding, and updating announcements. Created an Dto file for creating an announcement. Added a Announcement Model in the Prisma Schema and gave it a relationship with the coach. Imported the Announcement Module to the app.module.ts. (don't judge me on the naming of the branch I was nappy).