Created createClass. updateClass, and classResponse DTO.
createClass: deals with class creation. expects a class name, optional description, and an assignedCoachId. Also includes classDates which are represented by CreateClassDateDto

updateClass: updates the existing class. allows optional fields to be updated, such as class name, description and coach assigned to class. The classDates field also can be updated since it is also optional.

classResponse: is used when retrieving class data to return it as a response. returns class data from the service layer. it includes id, name, description, assignedCoachId, and an array of class dates.